# solutions/pocito/marketplace-server/marketplace-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "marketplace",
    "version": "0.1.0",
    "description": "Requests are scoped by the x-wonder-room header; omitting it uses the marketplace room."
  },
  "paths": {
...
    "/api/v1/tools/": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "List Resources",
        "description": "List every published tool config.",
        "operationId": "list_tools_api_v1_tools__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Tools Api V1 Tools  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Create Resource",
        "description": "Create a tool. 409 on id collision; 422 on invalid body.",
        "operationId": "create_tool_api_v1_tools__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTool"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/skills/": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "List Resources",
        "description": "List every published skill manifest.",
        "operationId": "list_skills_api_v1_skills__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Skills Api V1 Skills  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Create Resource",
        "description": "Create a skill. 409 on id collision; 422 on invalid body.",
        "operationId": "create_skill_api_v1_skills__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkill"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plugins/": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "List Resources",
        "description": "List every published plugin manifest.",
        "operationId": "list_plugins_api_v1_plugins__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Plugins Api V1 Plugins  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "plugins"
        ],
        "summary": "Create Resource",
        "description": "Create a plugin. 409 on id collision; 422 on invalid body.",
        "operationId": "create_plugin_api_v1_plugins__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlugin"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Resources",
        "description": "List every published agent manifest.",
        "operationId": "list_agents_api_v1_agents__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Agents Api V1 Agents  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create Resource",
        "description": "Create an agent. 409 on id collision; 422 on invalid body.",
        "operationId": "create_agent_api_v1_agents__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgent"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentConfig": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "system_prompt",
          "backend_config"
        ],
        "properties": {
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "backend_config": {
            "$ref": "#/components/schemas/BackendConfig"
          },
          "plugins": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "title": "Plugins"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "title": "Skills"
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "title": "Tools"
          },
          "sub_agents": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "title": "Sub Agents"
          },
          "knowledge_bases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledge Bases"
          }
        },
        "title": "AgentConfig",
        "description": "Agent payload with flat reference IDs and extensible backend-compatible fields."
      },
      "AssetInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "path",
          "content_b64"
        ],
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "content_b64": {
            "type": "string",
            "title": "Content B64"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          }
        },
        "title": "AssetInput",
        "description": "One asset attached to a skill. content_b64 is base64-encoded raw bytes."
      },
      "BackendConfig": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "harness_type"
        ],
        "properties": {
          "harness_type": {
            "type": "string",
            "enum": [
              "deepagents",
              "claude"
            ],
            "title": "Harness Type"
          }
        },
        "title": "BackendConfig",
        "description": "Backend harness selection with extensible fields for forward-compatible config.yaml files."
      },
      "CodeFile": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "path",
          "content"
        ],
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "title": "CodeFile",
        "description": "One already-wrapped tool source file stored byte-for-byte under tools/{id}/code/{path}."
      },
      "CreateAgent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "display_name",
          "description",
          "config"
        ],
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable semantic identifier used in API paths and references"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Human-readable display name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "hebrew_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hebrew Description"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagRef"
            },
            "title": "Tags"
          },
          "config": {
            "$ref": "#/components/schemas/AgentConfig"
          },
          "readme": {
            "type": "string",
            "default": "",
            "title": "Readme"
          }
        },
        "title": "CreateAgent",
        "description": "Create body. config.yaml is uploaded first; the manifest is written last."
      },
      "CreatePlugin": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "display_name",
          "description",
          "config"
        ],
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable semantic identifier used in API paths and references"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Human-readable display name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "hebrew_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hebrew Description"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagRef"
            },
            "title": "Tags"
          },
          "config": {
            "$ref": "#/components/schemas/PluginConfig"
          },
          "readme": {
            "type": "string",
            "default": "",
            "title": "Readme"
          }
        },
        "title": "CreatePlugin",
        "description": "Create body. config.yaml and README.md are uploaded first; the manifest is written last."
      },
      "CreateSkill": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "display_name",
          "description"
        ],
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable semantic identifier used in API paths and references"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Human-readable display name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "hebrew_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hebrew Description"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagRef"
            },
            "title": "Tags"
          },
          "min_agent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Agent Version"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "skill_md": {
            "type": "string",
            "default": "",
            "title": "Skill Md"
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetInput"
            },
            "title": "Assets"
          }
        },
        "title": "CreateSkill",
        "description": "Create body. SKILL.md and assets are uploaded first; the manifest is written last."
      },
      "CreateTool": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "display_name",
          "description",
          "tool_type"
        ],
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable semantic identifier used in API paths and references"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Human-readable display name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "hebrew_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hebrew Description"
          },
          "tool_type": {
            "$ref": "#/components/schemas/ToolType"
          },
          "json_schema": {
            "type": "object",
            "additionalProperties": true,
            "title": "Json Schema"
          },
          "is_async": {
            "type": "boolean",
            "default": true,
            "title": "Is Async"
          },
          "tracable": {
            "type": "boolean",
            "default": true,
            "title": "Tracable"
          },
          "dedicated_tool_config": {
            "type": "object",
            "additionalProperties": true,
            "title": "Dedicated Tool Config"
          },
          "code_files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CodeFile"
            },
            "title": "Code Files"
          }
        },
        "title": "CreateTool",
        "description": "Create body. Code files are uploaded before config.json; dedicated config is typed by tool_type."
      },
...
      "HTTPValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail"
          }
        },
        "title": "HTTPValidationError"
      },
      "PluginConfig": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "title": "Skills"
          },
        "tools": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Tools"
        },
        "knowledge_bases": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Knowledge Bases"
        }
      },
      "title": "PluginConfig",
      "description": "Flat resource IDs stored in config.yaml with forward-compatible extra fields."
      },
      "TagRef": {
        "type": "object",
        "required": [
          "tag_type",
          "tag_name"
        ],
        "properties": {
          "tag_type": {
            "type": "string",
            "title": "Tag Type"
          },
          "tag_name": {
            "type": "string",
            "title": "Tag Name"
          }
        },
        "title": "TagRef",
        "description": "A typed tag attached to a manifest."
      },
      "ToolType": {
        "type": "string",
        "enum": [
          "code",
          "flow_package",
          "flow_cube",
          "solr",
          "kick_graphql"
        ],
        "title": "ToolType",
        "description": "Invocation type: code uses a handler; dedicated runtime types use dedicated_tool_config."
      },
...
      "ValidationError": {
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "properties": {
          "loc": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "title": "ValidationError"
      },
...
    }
  }
}